Move non-core circuitpython files from root to lib
