+++ lib/commandsConfig.py
